feat: model
1.增加model层接口
2.增加container结构

+++ internal/model/basemodel.go
@@ -3,7 +3,7 @@ package model
 import "time"
 
 type BaseModel struct {
-	Id          int       `json:"id" gorm:"id"`
+	Id          int64     `json:"id" gorm:"id"`
 	CreatedTime time.Time `json:"created_time" gorm:"created_time"`
 	UpdatedTime time.Time `json:"updated_time" gorm:"updated_time"`
 	DeleteTime  time.Time `json:"delete_time" gorm:"delete_time"`
